[ads] Optimize segments

+++ components/brave_ads/core/internal/analytics/p2a/opportunities/p2a_opportunity.cc
@@ -12,7 +12,7 @@
 namespace brave_ads {
 
 namespace {
-constexpr int kTopSegmentsMaxCount = 1;
+constexpr size_t kTopSegmentsMaxCount = 1;
 }  // namespace
 
 void RecordP2AAdOpportunity(AdType ad_type, const SegmentList& segments) {
